+++ packages/core/src/components/search/UploadFileDropdown/UploadFileDropdown.tsx
@@ -461,6 +461,7 @@ export default function UploadFileDropdown({
             maxSize={MAX_FILE_SIZE}
             disabled={isProcessing}
             text={labels.dropzoneText}
+            aria-label={labels.dropzoneAriaLabel}
             dragActiveText={labels.dropzoneDragActiveText}
           />
           <Button